% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\title{Encode an AVIF image}
\usage{
encode(nr, quality = 80, speed = 10)
}
\arguments{
\item{nr}{A nativeRaster object}

\item{quality}{The quality of the image}

\item{speed}{The speed of processing}
}
\value{
A raw vector containing the encoded image data
}
\description{
Encodes an AVIF image from a nativeRaster object.
}
