% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{write_avif}
\alias{write_avif}
\title{Write an AVIF image to a file}
\usage{
write_avif(nr, path = NULL, quality = 80, speed = 10)
}
\arguments{
\item{nr}{A \code{nativeRaster} object.}

\item{path}{The path to write the image.}

\item{quality}{The quality of the image.}

\item{speed}{The processing speed.}
}
\value{
If \code{path} is \code{NULL}, returns a raw vector containing the encoded image data.
Otherwise, \code{path} is invisibly returned.
}
\description{
Write an AVIF image to a file
}
