% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rravif-package.R
\docType{package}
\name{rravif-package}
\alias{rravif}
\alias{rravif-package}
\title{rravif: AVIF Image Encoder}
\description{
Encodes images in AVIF format with 'ravif' Rust crate.
}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
